Fix dedup script
Forgot to update it to handle IFS correctly. Also return 0 in any case, the script shouldn't exit
with non-zero value in case it failed to create hardlinks, failure of this is non-fatal.

(I still have to figure out why it fails to create hardlinks inside the runner though, it seems weird)

Signed-off-by: Nick Kossifidis <mick@ics.forth.gr>